fix JISMatrix
fix writeJISToMatrixLED / writeJISToMatrixLEDs
(run with minus offset_from_left)

+++ src/JISMatrix.h
@@ -28,7 +28,7 @@ static inline size_t getJISRow(const char *utf8_s, uint8_t *out_buf, uint8_t row
  * @param c [in] Char to write.
  * @param offset_from_left [in] Offset size from left end.
 */
-size_t writeJISToMatrixLED(MatrixLED &matrixLED, const char *c, int8_t offset_from_left);
+size_t writeJISToMatrixLED(MatrixLED &matrixLED, const char *c, int16_t offset_from_left);
 
 /**
  * @brief Write JIS string to multiple matrixLED.
@@ -38,7 +38,7 @@ size_t writeJISToMatrixLED(MatrixLED &matrixLED, const char *c, int8_t offset_fr
  * @param offset_from_left [in] Offset size from left end.
  * @note Ascii char has 5px width, and JIS char has 7px. If you like monospace, you should NOT use ascii chars.
 */
-void writeJISsToMatrixLEDs(MatrixLED *matrixLEDs, uint8_t ledlen, const char *string, int8_t offset_from_left);
+void writeJISsToMatrixLEDs(MatrixLED *matrixLEDs, uint8_t ledlen, const char *string, int16_t offset_from_left);
 
 /* binarySearch in _UTF_CODES */
 const uint8_t *_binarySearchForJISMatrix(uint32_t target);
